+++ proto/src/lib.rs
@@ -24,4 +24,6 @@ pub mod oauth2;
 pub mod scim_v1;
 pub mod v1;
 
+pub mod attribute;
+
 pub use webauthn_rs_proto as webauthn;
