+++ proto/src/scim_v1/server.rs
@@ -1,3 +1,4 @@
+/*
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 use utoipa::ToSchema;
@@ -5,6 +6,7 @@ use utoipa::ToSchema;
 use scim_proto::ScimEntryHeader;
 
 use serde_json::Value as JsonValue;
+use crate::attribute::Attribute;
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, ToSchema)]
 pub struct ScimEntryGeneric {
@@ -13,3 +15,4 @@ pub struct ScimEntryGeneric {
     #[serde(flatten)]
     pub attrs: BTreeMap<Attribute, JsonValue>,
 }
+*/
